+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.1.0] - 2025-05-21
+
+### Added
+
+- IObservableVariableEventHolder interface for value to have events fired when
+value was stored or removed from a variable 
+
 ## [1.0.1] - 2025-03-18
 
 ### Added